ai-health-service/pyproject.toml
[tool.poetry]
name = "ai-health-service"
version = "1.0.0"
description = "A FastAPI-based service for uploading files to AWS S3, specifically designed for handling PDF documents in the AI Health ecosystem."
authors = ["AI Health Team <team@aihealth.com>"]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
boto3 = "^1.34.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
botocore = "^1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
start = "app.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"